+++ js/render/draw_circle.js
@@ -25,13 +25,9 @@ function drawCircles(painter, sourceCache, layer, coords) {
         const bufferGroups = bucket.bufferGroups.circle;
         if (!bufferGroups) continue;
 
-        const programOptions = bucket.paintAttributes.circle[layer.id];
-        const program = painter.useProgram(
-            'circle',
-            programOptions.defines,
-            programOptions.vertexPragmas,
-            programOptions.fragmentPragmas
-        );
+        const programConfiguration = bucket.programConfigurations.circle[layer.id];
+        const program = painter.useProgram('circle', programConfiguration);
+        programConfiguration.setUniforms(gl, program, layer, {zoom: painter.transform.zoom});
 
         if (layer.paint['circle-pitch-scale'] === 'map') {
             gl.uniform1i(program.u_scale_with_map, true);
@@ -51,8 +47,6 @@ function drawCircles(painter, sourceCache, layer, coords) {
             layer.paint['circle-translate'],
             layer.paint['circle-translate-anchor']
         ));
-
-        bucket.setUniforms(gl, 'circle', program, layer, {zoom: painter.transform.zoom});
 
         for (let k = 0; k < bufferGroups.length; k++) {
             const group = bufferGroups[k];
